dudas con el paso de props
