--- src/components/Modals.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import './styles/modals.css'
-
-const Modals = () =>{
-    return (<div className="container-cards">
-        <h1>Hola mundo</h1>
-    </div>);
-}
-
-export default Modals;
