+++ src/index.css
@@ -13,5 +13,5 @@ code {
 } */
 
 body {
-  background-color: black;
+  background-color: white;
 }
